Fix up import structure
