+++ opshin/rewrite/rewrite_inject_builtins.py
@@ -2,10 +2,9 @@ from copy import copy
 
 from ..typed_ast import *
 from ..util import (
-    PythonBuiltIn,
-    PythonBuiltInTypes,
     CompilingNodeTransformer,
 )
+from ..fun_impls import PythonBuiltIn, PythonBuiltInTypes
 
 """
 Inject initialising the builtin functions
